Changed for dedicated environment
